refactor: move the failure family into application/failures.py
