+++ tests/application/test_dependency_resolution.py
@@ -9,7 +9,7 @@ transitive propagation to dependents.
 
 from delta_engine.api import Column, DeltaTable, String
 from delta_engine.application.dependency_resolution import ResolveResult, resolve
-from delta_engine.application.results import ForeignKeyFailureReason
+from delta_engine.application.failures import ForeignKeyFailureReason
 from delta_engine.domain.model import QualifiedName
 from delta_engine.domain.model.foreign_key import ForeignKeyConstraint
 from delta_engine.domain.model.table import DesiredTable
