+++ tests/application/test_engine.py
@@ -3,23 +3,25 @@ import pytest
 from delta_engine.api import Column, DeltaTable, String
 from delta_engine.application.engine import Engine
 from delta_engine.application.errors import SyncFailedError
+from delta_engine.application.failures import (
+    ExecutionFailure,
+    ForeignKeyFailure,
+    ForeignKeyFailureReason,
+    ReadFailure,
+    ValidationFailure,
+)
 from delta_engine.application.registry import Registry
 from delta_engine.application.results import (
     CatalogState,
     ExecutionFailed,
-    ExecutionFailure,
     ExecutionResult,
     ExecutionSucceeded,
     ExecutionSummary,
-    ForeignKeyFailure,
-    ForeignKeyFailureReason,
     ReadFailed,
-    ReadFailure,
     SyncReport,
     TableAbsent,
     TablePresent,
     TableRunStatus,
-    ValidationFailure,
 )
 from delta_engine.domain.model import ObservedTable, QualifiedName
 from delta_engine.domain.model.foreign_key import ForeignKeyConstraint
